collectors/requirements.txt
pytrends>=4.9.0
keepa>=1.3.0
python-amazon-paapi>=6.0.0
APScheduler>=3.10.0
httpx>=0.25.0
python-telegram-bot>=20.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=4.9.0
pandas>=2.0.0
